+++ base/models/exam_enrollment.py
@@ -26,8 +26,9 @@
 from django.db import models
 from django.contrib import admin
 from django.utils.translation import ugettext_lazy as _
-from base.models import person, learning_unit_year
+from base.models import person, learning_unit_year, attribution, person_address, offer_year_calendar
 from django.utils import timezone
+import datetime
 
 
 class ExamEnrollmentAdmin(admin.ModelAdmin):
@@ -240,3 +241,125 @@ def find_for_score_encodings(session_exam_number,
     now = timezone.now()
     return queryset.filter(session_exam__offer_year_calendar__start_date__lte=now)\
                    .filter(session_exam__offer_year_calendar__end_date__gte=now)
+
+
+def scores_sheet_data(exam_enrollments, tutor=None):
+    exam_enrollments = sort_for_encodings(exam_enrollments)
+    data = {}
+    data['tutor_global_id'] = tutor.person.global_id if tutor else ''
+    now = datetime.datetime.now()
+    data['publication_date'] = '%s/%s/%s' % (now.day, now.month, now.year)
+    data['institution'] = str(_('ucl_denom_location'))
+    data['link_to_regulation'] = str(_('link_to_RGEE'))
+    # Will contain lists of examEnrollments splitted by learningUnitYear
+    enrollments_by_learn_unit = {}  # {<learning_unit_year_id> : [<ExamEnrollment>]}
+    for exam_enroll in exam_enrollments:
+        key = exam_enroll.session_exam.learning_unit_year.id
+        if key not in enrollments_by_learn_unit.keys():
+            enrollments_by_learn_unit[key] = [exam_enroll]
+        else:
+            enrollments_by_learn_unit[key].append(exam_enroll)
+
+    # # Sort by learningUnitYear.acronym then by Offeryear.acronym
+    # list_exam_enrollments = sorted(enrollments_by_learn_unit.values(),
+    #                                key=lambda k: "%s %s" % (k[0].session_exam.learning_unit_year.acronym,
+    #                                                         k[0].session_exam.offer_year_calendar.offer_year.acronym))
+    learning_unit_years =  []
+    for exam_enrollments in enrollments_by_learn_unit.values():
+        # exam_enrollments contains all ExamEnrollment for a learningUnitYear
+        learn_unit_year_dict = {}
+        # We can take the first element of the list 'exam_enrollments' to get the learning_unit_year
+        # because all exam_enrollments have the same learningUnitYear
+        session_exam = exam_enrollments[0].session_exam
+        learning_unit_year = session_exam.learning_unit_year
+        coordinator = attribution.find_responsible(learning_unit_year.learning_unit.id)
+        if coordinator:
+            coordinator_address = person_address.find_by_person_label(coordinator.person, 'PROFESSIONAL')
+
+        learn_unit_year_dict['academic_year'] = str(learning_unit_year.academic_year)
+        learn_unit_year_dict['coordinator'] = {'first_name': coordinator.person.first_name if coordinator else '',
+                                               'last_name': coordinator.person.last_name if coordinator else ''}
+        learn_unit_year_dict['coordinator']['address'] = {'location': coordinator_address.location
+                                                                      if coordinator_address else '',
+                                                          'postal_code': coordinator_address.postal_code
+                                                                         if coordinator_address else '',
+                                                          'city': coordinator_address.city
+                                                                  if coordinator_address else ''}
+        learn_unit_year_dict['session_number'] = session_exam.number_session
+        learn_unit_year_dict['acronym'] = learning_unit_year.acronym
+        learn_unit_year_dict['title'] = learning_unit_year.title
+        learn_unit_year_dict['decimal_scores'] = learning_unit_year.decimal_scores
+
+        programs = []
+
+        # Will contain lists of examEnrollments by offerYear (=Program)
+        enrollments_by_program = {}  # {<offer_year_id> : [<ExamEnrollment>]}
+        for exam_enroll in exam_enrollments:
+            key = exam_enroll.session_exam.offer_year_calendar.offer_year.id
+            if key not in enrollments_by_program.keys():
+                enrollments_by_program[key] = [exam_enroll]
+            else:
+                enrollments_by_program[key].append(exam_enroll)
+
+        for list_enrollments in enrollments_by_program.values():  # exam_enrollments by OfferYear
+            session_exam = list_enrollments[0].session_exam
+            offer_year = session_exam.offer_year_calendar.offer_year
+
+            deliberation_date = offer_year_calendar.find_deliberation_date(offer_year,
+                                                                               session_exam.number_session)
+            if deliberation_date:
+                deliberation_date = deliberation_date.strftime("%d/%m/%Y")
+            else:
+                deliberation_date = '-'
+            deadline = ""
+            if session_exam.offer_year_calendar.end_date:
+                deadline = session_exam.offer_year_calendar.end_date.strftime('%d/%m/%Y')
+
+            program = {'acronym': session_exam.offer_year_calendar.offer_year.acronym,
+                       'deliberation_date': deliberation_date,
+                       'deadline': deadline,
+                       'address': {'recipient': offer_year.recipient,
+                                   'location': offer_year.location,
+                                   'postal_code': offer_year.postal_code,
+                                   'city': offer_year.city,
+                                   'phone': offer_year.phone,
+                                   'fax': offer_year.fax,
+                                  }
+                       }
+            enrollments = []
+            for exam_enrol in list_enrollments:
+                student = exam_enrol.learning_unit_enrollment.student
+                enrollments.append({
+                    "registration_id": student.registration_id,
+                    "last_name": student.person.last_name,
+                    "first_name": student.person.first_name,
+                    "score": str(exam_enrol.score_final) if exam_enrol.score_final else '',
+                    "justification": exam_enrol.justification_final if exam_enrol.justification_final else ''
+                })
+            program['enrollments'] = enrollments
+            programs.append(program)
+        learn_unit_year_dict['programs'] = programs
+        learning_unit_years.append(learn_unit_year_dict)
+    data['learning_unit_years'] = learning_unit_years
+    return data
+
+
+def sort_for_encodings(exam_enrollments):
+    """
+    Sort the list by
+     1. offerYear.acronym
+     2. student.lastname
+     3. sutdent.firstname
+    :param exam_enrollments: List of examEnrollments to sort
+    :return:
+    """
+    def _sort(key):
+        off_enroll = key.learning_unit_enrollment.offer_enrollment
+        acronym = off_enroll.offer_year.acronym
+        last_name = off_enroll.student.person.last_name
+        first_name = off_enroll.student.person.first_name
+        return "%s %s %s" %(acronym if acronym else '',
+                            last_name.upper() if last_name else '',
+                            first_name.upper() if first_name else '')
+
+    return sorted(exam_enrollments, key=lambda k: _sort(k))
